requirements.txt
asyncio
paho-mqtt
requests
ecowater-softener
myPyllant
